trying to get data parsed correctly on the server side, and sent correctly on client side.
currently some of the data im parsing to JSON contains \, which are escape characters which stop it from sending correctly. need to find a way to nullify the esc chars before parseing through JSON.

also need to test multi client, and potential cross platform.

also need to investigate why i coded bytesize function to convert to str for the first loop, as this is messing with functions. wouldnt have done without a reason, but with all the changes, is this still the case, probably not.